Make all of the forms actually work
From the point where we moved to Rails, the forms weren't working
outside of automated tests. This is because we didn't have the CSRF
token set inside each form, but - unless you were paying close attention
- it looked like it had succeeded because the CSRF error triggered
an OAuth loop which ended in the user being shown the homepage, rather
than any kind of error.

By using the form_tag helper, we get the CSRF token (authenticity_token)
and UTF8 detection magic to boot.