setting name as email on zendesk ticket
For requests from users who are already in Zendesk,
a name is not necessary (just the email). Even if a name is provided
on the ticket, Zendesk ignores it. For users not yet in
Zendesk, a name must be provided on the ticket.

Since the forms no longer ask for the name (and very soon
the name will be coming from SSO), Zendesk rejects tickets from new
requesters. The stop-gap solution is to use the email address
as the name.