gracefully handling errors from Zendesk
Since Zendesk's API validations and error cases are no documented,
it is possible that the support app's validations let invalid requests
slip through. This means that it's not possible to raise a ticket
from the user's request.

Prior to this change, such cases would generate a standard Rails 500 error.
After this change, the user sees a much clearer error message and
receives a text representation of the ticket that was going to be
raised, in case they would like to review it or forward it on via email.