move contents to ticket description
before this change, the body of the ticket was created as a ticket comment
in Zendesk. This meant that the text was in the "ticket_audit" resource,
instead of the "ticket" resource in Zendesk world. Now it is the ticket
description.

This change brings the support app in line with the feedback app in this respect.