mock zendesk out at the HTTP level
up to now, the 'support' app has been tested using a dummy implementation
of the Zendesk API, coming from the `gds_zendesk` gem. This approach has allowed
bugs to creep in (because the dummy wasn't behaving in the same way as the real API).

This commit changes the integration tests to use 'webmock' to stub out the HTTP layer.
This change exposed several bugs (namely, that exceptions wouldn't be raised if something
went wrong on the Zendesk side), the fixes for which are also here.