don't raise tickets requested by suspended users
When a suspended user is set as the requester on a ticket, the Zendesk API
rejects the ticket creation. This change introduces a check to see whether
the user is suspended or not, and then raises or ignores the ticket.

This is a rehash of PR #86, which had to be reverted because of its
unfortunate side-effect of completely breaking the support app.

This one should actually work (honest).