Don't raise error when PP returns 404 on upload
This means that when PP returns 404, the data for that service
and that day won't be pushed into the PP (because retries won't occur).

The PP team will monitor for 404s and create buckets on their
end; the working assumption is that we don't need to capture that data
until the PP is set up to receive it, and it's stored in the contacts
DB anyway, so can be resent manually if this is really needed.