Feedex: limit the maximum number of results
Trying to show too many results in feedex will cause the app to time out.
This change introduces a limit (max 5000 results) to stop this from happening.
The limit is a somewhat arbitrary number, but should be enough results to provide
some insight but not enough to crash the app.

This limiting will become less relevant when filtering options (eg by date) become available.