Service feedback upload to PP: don't ignore 404s
This change reverts to the default gds-api-adapters behaviour
of blowing up when a 404 happens during an upload.

Previously, it was agreed with the PP team that when an upload failed
due to a 404 (ie the bucket was missing), it wasn't a problem that we
would lose the data for that day. Now, we would like to use the missing
data bucket exception as a prompt to go and create the bucket; also,
this means that sidekiq retries will ensure that no data is lost.