Add id to the Feedex JSON feeds
Right now, there isn't a way to 'freeze' the Feedex results to a certain
point in time. This causes problems when paginating over the results.

The long-term solution to this problem is to add a 'before' parameter
that's added to subsequent page links; however adding the record
'id' will allow to de-dupe records on the client-side if page-shift occurs.