Use support-api in FeedEx controller
That is, instead of querying the database directly.

Stub api requests in related specs, too.

We couldn't go as far as removing our ActiveRecord-backed models,
since there is still 'problem report metrics' code depending on it.

Until the necessary changes have been merged in gds-api-adapters and
support-api, this should not be merged!