WIP Get organisations list from Organisations API
To populate our upcoming 'Explore by Organisation' <select>

NOTE: This is flawed and will need replacing before this branch is merged. This
doesn't seem like it should be performed from a controller action. The current
implementation only actually fetches the first page of 20 orgs, and fetching all
of them seems like it might be slow.

Looking around at how other apps are using this, it seems Signon keeps its own
copy (of the subset of organisation data it needs) and uses a rake task to stay
in sync. Apparently the Need API provides a comparable service for its clients.

The approach for now will be to modify the Support API's Organisation model to
provide the attributes we need and fetch the list from there instead.