Use stubbed request for controller example
Just a case of one in a group of examples using the wrong `path` query parameter
for the request that's already stubbed for it within the setup for its group.

If RSpec always runs these specs in order and we're not clearing up after
creating our WebMock stubs, later specs will have access to earlier specs'
request stubs.