Add the ability to export global feedback
This section of the support application (feedex) now allows the
total quantity of requests received each day to be exported as a CSV
file.

Initially, we're protecting this behind a new role (`feedex_exporters`)
so we can control who gets access to the functionality. The main
reasoning here is to allow side-by-side verification of the figures
before rolling it out to wider (perhaps 'single point of contact')
release.