Ability to submit reports as reviewed for spam
A user can now submit a batch of problem reports as reviewed. Spam
checkboxes for marking as spam are wrapped in textless labels to
increase the target area in the UI for checking boxes. Each visible
checkbox element is accompanied by a hidden field, which makes it
possible to submit false as default to indicate that the report was
implicitly reviewed as not spam.