Use updated copy for titles and descriptions on forms
This change modifies the title of each form so that it matches the
request type. Additionally, the "guidance" or help text is displayed at
the top of the form. This is either some custom text or the "description" of
the request. I decided to remove "guidance" partials as I felt it's
clearer to have this small section of content inline in the form rather
than attempting to display the partial and catching render errors in the
outer view.