Configure emergency contact details by env var
We're moving this application to the newer deployment
pipeline so sensitive configuration should ideally be
stored in ENV vars.
In the case of the emergency contact details this is
a hash of different phone numbers and other info which
won't fit neatly into an env var as YAML.
So this has been stored as a JSON string and is parsed
accordingly by the application.