Get rails files to rails 4.2 standard
I ran `rake rails:update` to update existing files and generate new ones.
This means our app is now configured like a rails 4.2 app should be, and
this will make it easier to upgrade to rails 5.

This commit also migrates the app to use secrets.yml, and importantly
this means it uses the SECRET_KEY_BASE env var on deployed environments.
The infrastructure work to support this has already been done so we can
deploy this without further work. Prior to this we relied on the
secret_token.rb initializer being overwritten by our deployment process,
but that functionality was stopped in Dec 2016, meaning the app has been
insecure since then.