Add rails-controller-testing gem
In rails 5 the api for testing controllers changed drastically and parts
were extracted to their own gem.  In particular we use the `assigns`
method to check what objects were made available to the view after a
particular controller action was run.  We might want to rewrite these
specs to avoid using this helper as it's unlikely to remain in future
versions of rails or rspec.