Avoid missing template error when stubbing send_file
In rails 5 controller tests are full stack, and so the controller will
try to render a view.  The `send_file` api is the same as `render` in
that it acts to tell the framework that this action has completed and
rails doesn't need to fall through to the default render.  When we
stub it out to do nothing and return nil rails will then fall through
to the default render and complain about a missing template.  To avoid
this we provide a stub implementation to `send_file` that does `head :ok`
which is another `render` method that stops the framework from doing the
default render.  In this particular case we don't care about the actual
render value, so this works out ok for us.