Replace cancan with successor gem cancancan
The cancan gem stopped development in 2013, but the community forked it
and continued under the name cancancan.  It has API parity so we don't
need to change references to CanCan into CanCanCan.  This new gem has
the same version numbering too.  The main benefit here is that v. 1.16
brings full rails 5 compatibility by removing the dep warnings about
before_filters.