Require 'signin' permission from signon to access app
The expected behaviour of apps using signon is that they authenticate the
user and check that they have the 'signin' permission before allowing the
user to use the app.  This app only did the first part, not the second
bit.  The upside is that all signon users could access the support app;
the downside is that we have no way to limit a user so they can't use
this app, but still use other apps.  In practice this may not really be a
problem because of the purpose of the support app - it's unlikely we'd
want to allow someone to access other apps, but not support.  However it
does mean that the support app acts differently with respect to the
'signin' permission that is managed in signon and so this could be
confusing to future developers.

We will do some data work in signon to make sure that all existing users
retain signin permission, and that all new users are granted signin
permission for support by default.