Add path and organisation filter UI to feedback explorer
Before this commit we could only pick the path or organisation for
exploring feedback from the initial /explore URL.  It was possible to
change the path or org if you manually edited the URL, and to combine
them if you wanted.  This commit adds a UI for manipulating these filters
to the results page.  This lets you start from the /explore URL with a
organisation, and then latter add a path filter as well.