Make scope filters presenter extract paths from URIs
The UI suggests pasting in a full gov.uk URL and we'll extract the path
from it so we can't assume that the path will always be a path and not a
full URL. We extract the path from the value using the same algorithm as
`Scope::Requests::Anonymous::ExploreByUrl` which is used by the /explore
endpoint.

Now that much of the functionality is shared between the /explore endpoint
and the index action here we may want to consider removing the /explore
endpoint as even though we use ActiveModel validations on the Explore
objects to make sure we have valid form submissions, we don't actually let
the user know what the problems were if they fill in the form incorretly.