Manually fix ruby linter issues not automaticaly fixed
We tell the linter to not use the BlockLength cop on our tests because
it's not unusual to have long blocks when you use rspec which is entirely
block based.

We tell the linter to disable the FormatStringToken cop because `%{}` is
perfectly fine in our usecase of creating rails error messages - it's what
the docs tell you to do after all!

Other fixes are manually tidying up what the linter can't automatically
correct.