Fix scss linting errors
Mostly we're defining variables for colours and naming issues.  We have
to define two exclusions for the calendar.scss file though:

QualifyingElement - because we can't target the links with class
.ui-state-default without the element as they have no other attributes or
classes to identify them
PlaceholderInExtend - because we want to inherit styles from bootstrap in
some of our rules and these aren't written using placeholders so we can't
obey the rule

We disable these in the source, because scss-lint doesn't have the same
configuration cascade support that rubocop does so it's non-trivial for
us to use a local .scss-lint.yml config file and respect the config in the
govuk-lint gem.