Normalise the path when requesting a feedback export
We re-use the ScopesFilter object that the main feedback explorer
controller uses as this does the normalising that we need.  There's now
a fair bit of duplication between anonymous_feedback_controller and
anonymous_feedback/export_request_controller, but not enough to do
anything about it I don't think.  Will keep an eye on it though.