Refactor xxx_options methods in accounts permissions request
These serve two purposes:

1. to be passed as `collection:` params to forms to render the correct
   inputs and labels
2. to be used as lookups when rendering the formatted versions of the
   options when passing to zendesk

The 1st purpose is why they are in "human description" -> "computer value"
format as that's the way around that rails form helpers wants to render
things in collections.  The 2nd is why we turn things into a hash and then
look up by the value (short computer readable value) to get the key (the
long human readable description).  We can optimize this a little by storing
these options as hashes in the first place - rails form helpers will take
a hash for a collection as well as an array.  We can also store them on
the class and memoize them so we're not creating a new version each time
we call the method.

While we're at it we also try to use the hashes wherever we were referring
to the values or keys in it.  Specifically this means using the hash
values in our validations and predicate methods, and using the hash keys
in the views for form labels where we don't use the `collection` option.