Alphabetise gemfile (ish)
We keep rails at the top because it's the most important and we don't
want to lose it halfway down a big list.  We separately group the GDS
gems (although not in an actual bundler group, just their own stanza)
and we move the test group down below the development+test group.