Add note about why redis is an explicit dependency
We use redis directly to store user objects because the app has no access
to any other database to do so.  If we didn't do this we could remove the
explicit dependency and get it implicitly through our govuk_sidekiq
dependency, but it feels like it's worth calling this out explicitly lest
we forget.