Get brexit related feedback from the support api
As it stands there's a bug in the support app which makes it difficult to get feedback for a large number of endpoints at the same time that's quite a lot of work to fix. This is a relatively quick solution that pulls endpoints from google spreadsheets and compares the endpoints therein to feedback from the support api. I fully expect this to be deleted once that bug is fixed.