Pass personal data labels to Zendesk
Previously, content designers used a CSV file to tag content items to the business readiness finder.
In the CSV tagging process the label was the one presented to users in the facets, and the value was
Used by the csv to tag content.

In the new world content designers will be tagging content in Content Tagger, which will display the
label of the facet. To make it easier to tag content in Content Tagger, the Zendesk ticket should contain
the label, which is the same as the label that appears in content tagger.

The `personal_data_options` method was also used to display the checkboxes in the view, but as the label and value of
checkboxes can be the same, we can pass in the PERSONAL_DATA_OPTIONS list directly.