Allow users to add full url to request
Content-tagger allows content designers to use the full url to find
content to tag.
Also, it’s easier for the user to add a full URL and it lets 2nd line
click on the link from the ticket.

Update the test to find the id of the field rather than the text so that
this test doesn't need to be updated every time the question text
changes.