Fix exporting csv's with a large number of URL's
When users click 'Export as CSV' , they get the error '502 Bad Gateway' if
the URL has a large amount of paths present.

To solve this problem, instead of POST'ing the large number of URL's
we have added a path_set_id we get from Redis and POST the ID instead of the URL's.