Allow feedback exports with a raw string path parameter
Recently we refactored feedback CSV exports so that post requests are
sent with a Redis `path_set_id` rather than a large list of URLs.
https://github.com/alphagov/support/pull/672

Whilst it no longer possible with the Support app itself to access the AnonymousFeedbackController
without first generating a `path_set_id`, direct links do exist within the GOV.UK Chrome extension
and the Content-Data app. This causes the ExportRequestsController to send through a request to the
Support-API with an invalid `path_prefixes: scope_filters.paths_for_api`

This PR ensures that when directly accessing this controller with a string URL `path` or `paths` param,
clicking the `Export as CSV` button will cause the ExportRequestsController to create a valid
Redis `path_set_id` if one is not set.