Add additional account permission options
This commit adds the ability for a user to request permissions for 5
new applications, creating a new section in the `user_needs` partial.
It allows these as valid parameters to the controller, and if selected
will add the option to the formatted needs submitted to Zendesk.

As well as adding tests, it also removes a test for requesting account
access to Whitehall, as this is no longer done via Zendesk.