Add all Publishing services to the technical fault report page
Users can report technical faults on publishing services via the
technical fault report page. At the moment, there is an incomplete
list of services which can lead to confusion for users and colleagues
when triaging feedback.

This commit adds in missing Publishing services. It also adds in a
`do_not_know` option.

It felt wrong to add this to the list the `all` method in the
UserFacingComponents module, as it's not a user facing service.

I've chosen instead to update the technical_fault_report class to
add it as an extra option and ensure that validations are not
triggered if it selected by the user.

It might be a bit overkill, but it allows the UserFacingComponents class
to pass the correct user services back and increases reusability if this
application is ever extended.