Replace 'jc-validates_timeliness' with 'validates_timeliness'
The former is a fork of the latter, and was last updated in
September 2015: https://github.com/johncarney/validates_timeliness/commit/9e5e8f82e868748f8ca222e0e9904e158f630d6d
The original gem is still actively maintained.

We used to use the original, but switched to the fork in April
2015: https://github.com/alphagov/support/commit/9117b7a0c045e34dbcfe1e838368b270bb09646e

Since updating to Ruby 3.0, time validation is failing with errors
like:

```
Failures:

  1) Support::GDS::Campaign is expected to validate that :start_date cannot be empty/falsy
     Failure/Error: it { should validate_presence_of(:start_date) }

     ArgumentError:
       wrong number of arguments (given 3, expected 1..2)
     # ./spec/models/support/gds/campaign_spec.rb:30:in `block (2 levels) in <module:GDS>'
```

Presumably the gem is not compatible with the keyword argument
changes of Ruby 3.0:
https://rubyreferences.github.io/rubychanges/3.0.html#keyword-arguments-are-now-fully-separated-from-positional-arguments

Switching back to the original gem addresses some technical debt,
but we still see the same error. This appears to be because we're
now on Rails 7, which is not currently supported in the default
version of the gem. I've had to pin the version of the gem
to the 7.0 beta, which supports Rails 7 - see
https://github.com/adzap/validates_timeliness/issues/227.