Don't fall back to fixture data in production
The `EmergencyContactDetails` class relies on a `EMERGENCY_CONTACT_DETAILS`
ENV variable to define sensitive information such as phone numbers
and email addresses, which we then display in the Support app.

If the ENV variable is missing, then it falls back to the nonsense
data in `config/emergency_contact_details.json`, potentially
displaying incorrect information to publishers. It would be better
for the application to crash.

The fallback behaviour has instead been moved to
`config/environments`, in `development.rb` and `test.rb` only.
This is so that the app can be started up without having to define
the ENV var (`development.rb`) and so that all of the unrelated
tests continue to pass (`test.rb`).