Add Zendesk::CustomFieldType classes that format field value
Each class is responsible for validation and formatting of the value.

The naming in the Zendesk UI and API reference is inconsistent.  Decided to
follow the UI convention, as we are getting the values from a User support
agent who is using the UI [^1].  Types that are not currently used were not
included but the modules should be easy to extend.

We are currently using the following custom field types:
    - Text: default Zendesk type
    - Date: only the date (formatted as a ISO 8601 string) and omit the time.
      The form allows to submit dated in different valid formats. We need to
format and validate it, otherwise `ZendeskAPI::Error::RecordInvalid` will be
raised and the ticket will not be submitted. Since ‘Date’ class is already
defined on the main object, `DateField` is used.
    - Drop-down: the option's tag name for the value property, not the option
      text displayed in the list. We need to map the displayed text to a name
tag.

See [^2] for more examples and details.

[^1]: https://support.zendesk.com/hc/en-us/articles/4408838961562
[^2]: https://developer.zendesk.com/documentation/ticketing/managing-tickets/creating-and-updating-tickets/#setting-custom-field-values