Add Zendesk::CustomField class
We want to start populating custom fields in Zendesk.

To set the value of one or more custom fields in a ticket, we need to specify
an array of objects consisting of `id` and `value` properties. [^1]

* ID integers
Most Zendesk Support resources, such as tickets and users, are identified by
the integer specified by the id attribute of API responses. The term refers to
the JSON numeric type, not a C-like int data type.  The default numeric type in
JavaScript, Ruby, Python, and PHP is sufficient to represent Zendesk Support ID
integers.
Tested that String also work as an ID but decided to follow Zendesk
documentation on this.
ID formatting curtesy of rubocop:
   Style/NumericLiterals: Use underscores(_) as thousands separator and
separate every 3 digits with them.
(https://rubystyle.guide#underscores-in-numerics)

* Value
Value depends on the custom field type. Validation and formatting is preformed
by each class.

This particular implementation (reading ruby classes from a YAML file)
introduces a level of abstraction but has the following benefits:
- allows us to store all information about the Zendesk Custom fields in a YAML
  file (the app doesn't seem to have access to the custom fields end point via
Zendesk API)
- specific Ticket classes don't need to know about the type of field
- specific Ticket classes don't need to know about field type
- CustomField class is not responsible for validating/formatting values of
  different types
- makes it easy to extend should we want to populate other type of fields such
  as multi-select.

[^1]: https://developer.zendesk.com/documentation/ticketing/managing-tickets/creating-and-updating-tickets/#setting-custom-field-values