Don't duplicate info in the Content Request ticket body
The information is included in the custom fields so this removes duplicate
information as requested by a Content Designer.  It removes the
base_comment_snippets from ZendeskTicket. The specific Ticket classes can
define whether the information is passed in the body or custom fields.  It also
moves the [Reason for time constraint] at the end of the ticket body.