Send Content change request tickets to "Content Request" form
Normally the Zendesk tickets default to the "Default Ticket Form. We want to
render "Content Request" form because some information in the Content change
request tickets is included only in the custom fields (not in the body). The
Zendesk user might not see this information if they don’t know to switch to the
correct form template. Therefore we want to render the correct form for this
ticket.

Name: ticket_form_id
Type: integer
Read-only: false
Mandatory: false
Description: Enterprise only. The id of the ticket form to render for the ticket

From the API reference: https://developer.zendesk.com/api-reference/ticketing/tickets/tickets/#json-format