Temporarily add content request fields into the ticket body
There is an issue with Zapier integration the content teams use to create
Trello cards from Zendesk tickets.  Deadlines/not-before dates aren’t pulling
through on cards because this data is now in different fields. This was
anticipated however colleagues with with login details to the Zapier account
are not available today to make a change.  We will temporally add this
information to the ticket body so it's copied over to Trello with Zapier.

Once Zapier is updated we can revert this commit.