Remove broken JS for toggling requested user fields
This JS was introduced in this PR [1] to hide the job & phone fields on
the form for what was then the CreateOrChangeUserRequest. However, the
JS relied on the input IDs that the Rails form builder generated and
thus it broke when CreateOrChangeUserRequest was renamed to
AccountsPermissionsAndTrainingRequest in this commit [2] some 8 years
ago. Since we're planning to split out a separate request workflow for
changing an existing user, this doesn't seem like a big loss!

[1]: https://github.com/alphagov/support/pull/72
[2]: https://github.com/alphagov/support/commit/0ea452316139b1e49e7e6f3d9f4f4bbfbef7316f

Co-authored-by: Chris Roos <chris@seagul.co.uk>