Copy new/change user workflow -> change user workflow
Trello: https://trello.com/c/aQiPdsiP

We want to simplify the "Accounts, permissions and training" request
workflow by splitting it into two workflows - one for creating a new
user and one for changing an existing user. As a first step we've
*copied* the AccountsPermissionsAndTrainingRequest class and all its
associated classes to a new ChangeExistingUserRequest class. We're
planning on renaming the AccountsPermissionsAndTrainingRequest class and
its associated classes to CreateNewUserRequest in a subsequent commit.

Co-authored-by: Chris Roos <chris@seagul.co.uk>