Add organisation to requested user fields for new user request
Trello: https://trello.com/c/aQiPdsiP

To match the new designs.

The field is intentionally not a required field, because we need to
cater for local authority users.

We've borrowed a bunch of the code for this from
AnonymousFeedback::ExploreController. However, we want to record the
human-friendly organisation title + acronym rather than the organisation
slug, so we're mapping from Organisation objects to select options
slightly differently.

I've decided to add organisation as an attribute of RequestedUser rather
than of CreateNewUserRequest, because that made the most sense to me. It
does mean we're now making it available to the Zendesk API client.
However, the latter will just ignore it. Also we're not capturing the
organisation in ChangeExistingUserRequest which also uses RequestedUser,
but again this shouldn't matter.