Fix layout of organisation select on new user request form
Previously the organisation select element was on the same line as its
label; whereas the name & email text inputs were below their labels.

The select2 JS library constructs DOM elements to replace the standard
select input. This constructed element doesn't retain the `input-md-7`
class which means that it ends up alongside the "Organisation" label.

This commit moves the whole form into a row whose content has a maximum width
of 6 columns in the bootstrap grid. This forces the select element onto
the next line in all page widths and we no longer need the `input-md-*`
CSS classes on the individual form controls.

This means the layout now more closely matches the new designs.

This diff is best viewed using the `--ignore-all-space` git option,
because of the indentation changes.