Allow select2-ified select elements  to be cleared
Since the organisation select for CreateNewUserRequest is not required,
it feels as if the user ought to be able to clear it if they
accidentally select an organisation.

I've applied these options *globally*, so they affect the drop-downs for
"Organisation" & "Document Type" on a couple of other pages (i.e.
`/anonymous_feedback/explore` & `/anonymous_feedback`). I can't see why
making any of these dropdowns clearable would be a problem and the empty
string placeholder seems to make sense in all cases, because they all
pass the `allow_blank: true` option to
`ActionView/Helpers/FormOptionsHelper#options_for_select`.