Fix NoMethodError in create user form when validation fails
Previously when server-side form validation failed, an attempt was made
to re-render the form and display the errors. However, since the
`organisations_list` instance variable wasn't set in the `create`
action, a NoMethodError was raised within the call to
`#options_for_select` in the template.

The controller inheritance makes it awkward to set the
`organisations_list` instance variable at the appropriate point in the
`create` action, so instead I've opted to make the organisation options
available from a controller helper method which is accessible from the
template.

Fortunately this probably hasn't affected many users, because most/all
of the validations are also covered by client-side validation which is
supported in most browsers [1].

[1]: https://caniuse.com/mdn-html_elements_form_novalidate