Simplify RequestedUser by using latest gds_zendesk
Previously the reqested_user argument passed to
GDSZendesk::Users#create_or_update_user had to respond to both `#job`
and `#phone` methods so that it could set the `details` and `phone` on
the Zendesk user it created or updated.

However, since gds_zendesk v3.6.0 the requested_user no longer has to
respond to those methods, so we can simplify Support::GDS::RequestedUser
accordingly.