Move requested user attributes into parent request
Only two request types, CreateNewUserRequest &
ChangeExistingUserRequest, were using these attributes and while
separating them out into nested attributes avoided a bit of duplication,
it came at a cost of quite a bit of extra complexity.

Unlike CreateNewUserRequest, ChangeExistingUserRequest doesn't have a
form field for the requested user's organisation, which meant the code
was a bit confusing.

Moving the validation for these attributes onto the parent request means
that we can provide more helpful error messages. Previously we were just
displaying "The details of the user in question are either incomplete or
invalid" as the error message, rather than e.g. "Name can't be blank",
"Email is invalid", etc. This is about to become more important, because
we're about to move these two forms to use the GOV.UK Design System and
so we won't be able to rely on client-side validation and we need to
display errors against individual fields as well as in a summary.