Avoid two email validation errors when email is blank
Now that we're displaying *per-attribute* validation errors for the
create/change user request forms it would be confusing to see two
validation errors for the `email` attribute when the attribute is blank:
one triggered by the `presence` validation; and one triggered by the
`format` validation.

This adds the `allow_blank` option to the `format` validation, so that
now that validation is only triggered if the `email` attribute is not
blank.