Add platforms to Gemfile.lock
c.f. https://github.com/alphagov/whitehall/pull/6934

We ran the following commands:

    bundle lock --add-platform ruby
    bundle lock --add-platform aarch64-linux
    bundle lock --add-platform arm64-darwin
    bundle lock --add-platform x86_64-linux

We've been seeing discrepancies in the Gemfile.lock when adding new gems
to this app whereby the new gem only gets added for the current bundler
platform. We're hoping that making this change which has been made in a
bunch of other GOV.UK apps will avoid that happening in future PRs.

Co-authored-by: Chris Roos <chris@seagul.co.uk>