Use GOV.UK Design System for create new user request form
Trello: https://trello.com/c/aQiPdsiP

* Use the recently added "design_system" layout and the
`GOVUKDesignSystemFormBuilder::FormBuilder` in the
`CreateNewUserRequestsController` so the design system changes are
restricted to this one form.

* Modify the `_request_details.html.erb` template and add an overriding
`_collaborators.html.erb` template to use the new design and form
builder for all the form elements.

* Add an overriding `new.html.erb` to make use of the
`#govuk_error_summary` and `#govuk_submit` form builder methods. I plan
to make other changes in this template in subsequent commit which I
think justify the small amount of duplication introduced at this stage.

* Add a `content_for(:error_summary)` to the `design_system` layout so
that it appears in the right place on the page.

* Add a blank option to the controller `#organisation_options` method,
because the `#govuk_collection_select` form builder method doesn't
support an `allow_blank` option.

* Modify and enhance tests for validation errors to highlight that
errors are now displayed inline against the relevant form field as
well as in an error summary at the top of the page.

Note that we've lost the functionality provided by `select2` which
previously allowed users with JavaScript enabled to search for
organisations in the drop down. I plan to address that in a subsequent
commit.

Also note that there are no breadcrums in the new "design_system"
layout. However, according [1] to the GOV.UK Design System, there is no
need for them on this page, because it is a "top-level" page and its
parent page, the home page, is available in the navigation links.

[1]: https://design-system.service.gov.uk/components/breadcrumbs/